// projects/projects/src/api/model/content.ts
/**
 * MaestroHost - SceneService
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Scene } from './scene';
import { Target } from './target';

export interface Content {
  contentId?: number;
  name?: string;
  description?: string;
  scene?: Scene;
  file?: any;
  numberOfVisits?: number;
  createdAtUTC?: Date;
  inUse?: boolean;
  typeId?: number;
  targets?: Array<Target>;
  readonly sceneId?: number;
}